Split up base Manticore tests and logging tests
The verbosity changes seem to be taking hold when they shouldn't
