+++ tests/native/test_logging.py
@@ -1,0 +1,21 @@
+import unittest
+import logging
+
+from manticore.utils.log import get_verbosity, set_verbosity
+
+
+class ManticoreLogger(unittest.TestCase):
+    """Make sure we set the logging levels correctly"""
+
+    _multiprocess_can_split_ = True
+
+    def test_logging(self):
+        set_verbosity(5)
+        self.assertEqual(get_verbosity("manticore.native.cpu.abstractcpu"), logging.DEBUG)
+        self.assertEqual(get_verbosity("manticore.ethereum.abi"), logging.DEBUG)
+
+        set_verbosity(1)
+        self.assertEqual(get_verbosity("manticore.native.cpu.abstractcpu"), logging.WARNING)
+        self.assertEqual(get_verbosity("manticore.ethereum.abi"), logging.INFO)
+
+        set_verbosity(0)
