+++ tests/native/test_manticore.py
@@ -134,20 +134,3 @@ class ManticoreTest(unittest.TestCase):
         else:
             self.assertTrue(a <= 0x41)
             self.assertTrue(b > 0x41)
-
-
-class ManticoreLogger(unittest.TestCase):
-    """Make sure we set the logging levels correctly"""
-
-    _multiprocess_can_split_ = True
-
-    def test_logging(self):
-        set_verbosity(5)
-        self.assertEqual(get_verbosity("manticore.native.cpu.abstractcpu"), logging.DEBUG)
-        self.assertEqual(get_verbosity("manticore.ethereum.abi"), logging.DEBUG)
-
-        set_verbosity(1)
-        self.assertEqual(get_verbosity("manticore.native.cpu.abstractcpu"), logging.WARNING)
-        self.assertEqual(get_verbosity("manticore.ethereum.abi"), logging.INFO)
-
-        set_verbosity(0)
